requirements.txt
langgraph
langchain
langchain-openai
langchain-community
yfinance
tavily-python
python-dotenv
pandas
matplotlib
